Common : Fixed the timeGetTime() method on POSIX systems. Added a atomic_inc() wrapper over GCC / Visual Studio. Made all singletons' getInstance() methods thread-safe.
Database Module :
Made some database functions constant.
Updated the createPlayer() function.
Updated all queries to use enum for result positions.
Updated most queries to convert to UInt() instead of Int().
Added support for CPs to the player.
Added a method to get the required exp of a level.
Added a method to get the required uplvl times of a level.
Added a method to get the points distribution of a profession / level.
Added a method to load all NPCs' tasks.
Added a method to load all monsters.
Added a method to load all generators.
Moved the items to an internal map.

Entities Module :
Updated the broadcast set to be thread-safe.
Updated the monster class to work with the generators.
Added support for CPs to the player.
Updated the move() method to check for the elevation.
Added a kickBack() method to the Player class.
Updated the Player class to remove the player from the map at destruction.
Added constants for SEX, LOOK, FACE, HAIR, PKMODE to Player.
Updated the exp variable to be 64 bits instead of 32 bits.

Map Module :
Updated the GameMap class to be thread-safe when using the entities map.
Updated the MapData pack system with references.
Blocked the MapData pack system until a fix is found.

MsgAction :
Updated the CHG_MAP action to kickBack() properly.
Finished the SET_PKMODE action to update the Pk Mode.
Added the support for JUMP action.

MsgPlayer :
Updated the structure.

MsgTick :
Updated the structure and the implementation.

Generator Module :
Added the generator implementation based on TQ's one.

World Module :
Updated the world module to use QtConcurrent instead of threads.
Updated the world to be thread-safe when accessing players' maps.
Added a worker to regenerate monsters.
Added a worker to process players.