Project : Simplified the Qt version detection. Added NDEBUG flag when building in release.
Common :
Updated the SAFE_DELETE macros to work on a single line.
Updated the ASSERT macro to be executed in release too.
Added a constexpr macro (for future use).
Moved the CLOCK_MULTIPLIER constant to basefunc.cpp.
Fixed the generator (still generating when the map was inactive).

Server :
Changed the Server implementation from Singleton pattern to QCoreApplication.

Database :
Added a validation of the SQL driver for the LastInsertId feature.
Added the getPlayerItems() method in the database.
Added the saveItem() method in the database.
Added the createItem() method in the database.
Added the eraseItem() method in the database.
Added the getItemFromShop() method in the database.
Updated the MySQL dump file.

Scripts :
Added Conductress scripts.
Added old KCOSS scripts as reference.
Added a Lua method to get the Mate.
Added Lua methods to add/spend money/CPs.
Added a Lua method to move the player.
Merged ItemTask & NpcTask to the Script class.

Entities :
Removed the time parameter of timerElapsed().
When activating a NPC, set the task to the client.

Player :
Added a timer for adding energy to the player.
Added methods for adding/spending money/CPs to the player.
Added a method for adding a value to an attribute to the player.
Added a method to find an item in the inventory.
Added a method to award an item.
Added a method to add an item to the inventory.
Added a method to erase an item from the inventory.
Added a method to send the info of all items.
Added a method to save all items.
Added a method to free all items.
Added methods to determine what is the item and what can be done.

Item :
Added a factory method to create an item and insert it in the database.
Added a factory method to create an item from a Sql query.
Added a method to save the item.
Added a method to erase the item.
Added a method to get the sell price of the item.
Added a method to get the repair cost of the item.

Msg / Networking :
Loading items at connection (MsgConnect).
Sending items at login (ACTION_GET_ITEM_SET).
Fixed the warning of hidden virtual function in MsgLoginChallengeS.
Fixed the command parsing in MsgTalk to handle white-spaces.
Fixed and finalized the MsgUserAttrib packet.
Fixed and finalized the MsgItemInfo packet.
Added support for buying, selling & repairing items.
Added basic support for executing the next Idx of a NPC's task.