Server: Scripts are now loaded before items & NPCs.
Database:
Added missing data to Item::Info.
Added tasks' loading for items.
Updated tasks' loading for NPCs.
Moved Lua scripts to tasks folder.

World:
Added a method to query a task.

Player:
Added a method for determining if an item can be equipped.
Added a method to use an item (and equip).
Added a method to equip an item.
Added a method to unequip an item.
Possibly fixed the addItem() method.

Item:
Added missing data to Item::Info.
Added a constructor to avoid invalid values.
Added documentation to the class.
Added a method to set the position.
Added methods to determine what is the item.

Msg / Networking :
Fixed the structure of the MsgItem packet.
Added support for using item.
Fixed the swapping of MsgNpc packet.